requirements.txt
gradio
transformers
torch
sentencepiece
sacremoses
PyPDF2
python-docx
nltk
reportlab